- Implemented checks in delete_note, get_notes, and save_note to ensu…

+++ classes/external/delete_note.php
@@ -71,6 +71,10 @@ class delete_note extends \core_external\external_api {
         $context = context_course::instance($course->id);
         self::validate_context($context);
 
+        if (!\local_quicknote\hooks::is_enabled_for_course($course)) {
+            throw new \moodle_exception('disabledforcourse', 'local_quicknote');
+        }
+
         $DB->delete_records('local_quicknote_notes', ['id' => $note->id]);
 
         return [
